Ensure that CVScan waits until scan is complete
If the beam is down when a cvscan starts, the moveTo method would return
before the controller was started. This caused the waitWhileBusy method
to return immediately and the scan to then fail as no file had been
written by the end of the scan.

The actual scan then runs in the background after GDA returns to an idle
state.

This explicitly sets the controller to busy if the beam is down at the
start of a scan. As the busy state of the controller is not relied on
internally, it should not affect the running of the scan.

See I11-583

Change-Id: I4dc24ea4380ec434f82e85f870010a53c653e110